move svc interception to a dedicated file
