+++ src/emulate.h
@@ -18,16 +18,7 @@
 
 #include "cx_crc.h"
 
-typedef enum {
-  SDK_NANO_X_1_2,
-  SDK_NANO_S_1_5,
-  SDK_NANO_S_1_6,
-  SDK_BLUE_2_2_5,
-  SDK_LAST
-} sdk_version_t;
-
-// Max number of characters allowed in version number
-#define MAX_SDK_VER_LEN  6
+#include "sdk.h"
 
 /* TODO */
 typedef void bolos_ux_params_t;
